third_party/WebKit/LayoutTests/fast/css/currentcolor-inherit.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<style>
#outer {
    background-color: currentColor;
    color: red;
}
#inner {
    background-color: inherit;
    border-color: inherit;
    color: green;
    height: 100px;
}
</style>

<div id=outer>
    <div id=inner></div>
</div>

<script>
test(function() {
    assert_equals(getComputedStyle(inner).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(inner).borderColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(inner).borderLeftColor, "rgb(0, 128, 0)");
}, "currentColor should be inherited unresolved");
</script>
